Cut the page down to one screen
The forecast chart goes, and with it the scroll. StatewideHero and its
stylesheet are deleted outright, useSize with them — it had no other caller.
d3-scale and d3-shape stay; the sparkline uses them now. Nothing follows the
stage, so scrollHeight equals the viewport exactly and there is nowhere to
scroll to.

The chart's key survives it, since the sparkline carries the same two bands:
it sits on the Run button's row, right-aligned under the date and time, so it
costs no height of its own. Labelled "Forecast", not "actual" — actuals are
observed CAISO output, a series this page does not carry at all.

Weekdays are spelled out.

One thing the chart was quietly carrying: the standing caveat about what is
graded and what is estimated, which WEB.md makes a non-negotiable of this
product. With no space below the fold it now runs as a small line along the
bottom of the deck. Cutting it entirely would have been the easy reading of
"remove the chart", and it is the one part I did not remove.

+++ src/lib/time.ts
@@ -3,16 +3,16 @@
  * arithmetic in it — differences are taken on epoch milliseconds, and the
  * time zone only ever enters at the formatting step.
  *
- * Clocks read as 12-hour with AM/PM. The date half stays day-month ("Sat
- * 15 Jun"), so the two are composed from separate formatters rather than
- * pulled out of one locale's idea of how to join them.
+ * Clocks read as 12-hour with AM/PM. The date half stays day-month
+ * ("Saturday 15 Jun"), so the two are composed from separate formatters
+ * rather than pulled out of one locale's idea of how to join them.
  */
 
 const PT = 'America/Los_Angeles';
 
 const dateOnly = new Intl.DateTimeFormat('en-GB', {
   timeZone: PT,
-  weekday: 'short',
+  weekday: 'long',
   day: 'numeric',
   month: 'short',
 });
@@ -44,12 +44,12 @@ const hour24 = new Intl.DateTimeFormat('en-US', {
   hour12: false,
 });
 
-/** "Sat 15 Jun, 11:00 AM PT" */
+/** "Saturday 15 Jun, 11:00 AM PT" */
 export function formatPacific(iso: string): string {
   return `${formatPacificDate(iso)}, ${formatPacificTime(iso)} PT`;
 }
 
-/** "Sat 15 Jun" */
+/** "Saturday 15 Jun" */
 export function formatPacificDate(iso: string): string {
   return dateOnly.format(new Date(iso));
 }
